FP1-7 long to integer

+++ Backend/src/main/java/com/example/data/BankCard.java
@@ -14,7 +14,7 @@ import javax.persistence.Table;
 @Table(name = "bank_card")
 public class BankCard {
 
-	public BankCard(Integer cardNumber, String sortCode, Integer bankAccountId, Integer accountNumber, String bank) {
+	public BankCard(Long cardNumber, String sortCode, Integer bankAccountId, Integer accountNumber, String bank) {
 
 		super();
 		this.cardNumber = cardNumber;
@@ -24,8 +24,8 @@ public class BankCard {
 
 	}
 
-	public BankCard(Integer bankCardId, Integer cardNumber, String sortCode, Integer bankAccountId,
-			Integer accountNumber, String bank) {
+	public BankCard(Integer bankCardId, Long cardNumber, String sortCode, Integer bankAccountId, Integer accountNumber,
+			String bank) {
 
 		super();
 		this.bankCardId = bankCardId;
@@ -45,7 +45,7 @@ public class BankCard {
 	private Integer bankCardId;
 
 	@Column(name = "card_number")
-	private Integer cardNumber;
+	private Long cardNumber;
 
 	@Column(name = "sort_code")
 	private String sortCode;
@@ -74,11 +74,11 @@ public class BankCard {
 		this.bankCardId = bankCardId;
 	}
 
-	public Integer getCardNumber() {
+	public Long getCardNumber() {
 		return cardNumber;
 	}
 
-	public void setCardNumber(Integer cardNumber) {
+	public void setCardNumber(Long cardNumber) {
 		this.cardNumber = cardNumber;
 	}
